requirements.txt
yamltools
jsonpath-ng==1.4.3
jsonschema==2.6.0
PyYAML==3.12
